remove extra load_dotenv
