+++ src/models/database.py
@@ -1,11 +1,7 @@
 import os
-from typing import Generator
 
-from dotenv import load_dotenv
 from sqlalchemy import create_engine
-from sqlalchemy.orm import Session, declarative_base, sessionmaker
-
-load_dotenv()
+from sqlalchemy.orm import declarative_base, sessionmaker
 
 SQLALCHEMY_DATABASE_URL = os.getenv("SQLALCHEMY_DATABASE_URL")
 
@@ -14,14 +10,3 @@ engine = create_engine(SQLALCHEMY_DATABASE_URL)
 SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
 
 Base = declarative_base()
-
-
-def get_db() -> Generator[Session, None, None]:
-    """
-    A generetor function that yields the DB session
-    """
-    try:
-        db = SessionLocal()
-        yield db
-    finally:
-        db.close()
