+++ tests/client.py
@@ -1,16 +1,12 @@
 import os
 from typing import Generator
 
-from dotenv import load_dotenv
 from fastapi.testclient import TestClient
 from sqlalchemy import create_engine
 from sqlalchemy.orm import Session, sessionmaker
 
 from main import app
-from src.models.database import get_db
-
-load_dotenv()
-
+from src.dependencies import get_db
 
 engine = create_engine(os.getenv("SQLALCHEMY_DATABASE_URL_TEST"))
 TestingSessionLocal = sessionmaker(
